refactor: delete unused types, change structure project
