+++ src/interfaces/IAlbum.ts
