+++ src/interfaces/IArtist.ts
