+++ src/interfaces/IFavorites.ts
