+++ src/tracks/tracks.controller.ts
@@ -18,33 +18,30 @@ export class TracksController {
   constructor(private readonly TracksService: TracksService) {}
 
   @Get()
-  async getAll() {
+  getAll() {
     return this.TracksService.getAll();
   }
 
   @Get(':id')
-  async getById(@Param('id') id: string) {
+  getById(@Param('id') id: string) {
     return this.TracksService.getById(id);
   }
 
   @Post()
   @UsePipes(ValidationPipe)
-  async createTrack(@Body() CreateTrackDto: CreateTrackDto) {
+  createTrack(@Body() CreateTrackDto: CreateTrackDto) {
     return this.TracksService.createTrack(CreateTrackDto);
   }
 
   @Put(':id')
   @UsePipes(ValidationPipe)
-  async updateTrack(
-    @Body() UpdateTrackDto: CreateTrackDto,
-    @Param('id') id: string,
-  ) {
+  updateTrack(@Body() UpdateTrackDto: CreateTrackDto, @Param('id') id: string) {
     return this.TracksService.updateTrack(UpdateTrackDto, id);
   }
 
   @Delete(':id')
   @HttpCode(204)
-  async remove(@Param('id') id: string) {
+  remove(@Param('id') id: string) {
     return this.TracksService.deleteTrack(id);
   }
 }
