+++ src/users/users.controller.ts
@@ -23,24 +23,24 @@ export class UsersController {
   constructor(private readonly UsersService: UsersService) {}
 
   @Get()
-  async getAll() {
+  getAll() {
     return this.UsersService.getAll();
   }
 
   @Get(':id')
-  async getById(@Param('id') id: string) {
+  getById(@Param('id') id: string) {
     return this.UsersService.getById(id);
   }
 
   @Post()
   @UsePipes(ValidationPipe)
-  async createUser(@Body() CreateUserDto: CreateUserDto): Promise<UserEntity> {
+  createUser(@Body() CreateUserDto: CreateUserDto) {
     return new UserEntity(this.UsersService.createUser(CreateUserDto));
   }
 
   @Put(':id')
   @UsePipes(ValidationPipe)
-  async updateUser(
+  updateUser(
     @Body() UpdatePasswordDto: UpdatePasswordDto,
     @Param('id') id: string,
   ) {
@@ -49,7 +49,7 @@ export class UsersController {
 
   @Delete(':id')
   @HttpCode(204)
-  async remove(@Param('id') id: string) {
+  remove(@Param('id') id: string) {
     return this.UsersService.deleteUser(id);
   }
 }
